+++ package/src/@types/event.ts
@@ -1,37 +1,94 @@
 /**
- * Kind of the copied item.
+ * Kind of the resolved target.
  */
-type CopyTargetKind = "file" | "directory";
+type ResolvedTargetKind = "file" | "directory";
 
 /**
- * Event emitted for each copied item.
+ * Resolved target.
  */
-type CopyEvent = {
+type ResolvedTarget = {
     /**
      * Kind of the copied item.
      */
-    kind: CopyTargetKind;
+    kind: ResolvedTargetKind;
     /**
-     * Absolute source path of the copied item.
+     * Absolute source path of the target.
      */
     src: string;
     /**
-     * Absolute destination path of the copied item.
+     * Absolute destination path of target.
      */
     dest: string;
     /**
-     * Whether the item was renamed.
+     * Whether the target was renamed.
      */
     renamed: boolean;
     /**
-     * Whether the item was transformed.
+     * Whether the target was transformed.
      */
     transformed: boolean;
 };
 
 /**
- * Listener for copy events.
+ * Copy start event.
  */
-type CopyEventListener = (event: CopyEvent) => void | Promise<void>;
+type CopyStartEvent = {
+    /**
+     * Current working directory.
+     */
+    cwd: string;
+    /**
+     * Resolved targets.
+     */
+    targets: ResolvedTarget[];
+};
 
-export type { CopyEvent, CopyEventListener, CopyTargetKind };
+/**
+ * Listener for copy start event.
+ */
+type CopyStartEventListener = (event: CopyStartEvent) => void | Promise<void>;
+
+/**
+ * Copy event.
+ */
+type CopyEvent = {
+    /**
+     * Resolved target.
+     */
+    target: ResolvedTarget;
+};
+
+/**
+ * Listener for copy event.
+ */
+type CopyEventListener = (context: CopyEvent) => void | Promise<void>;
+
+/**
+ * Copy end event.
+ */
+type CopyEndEvent = {
+    /**
+     * Current working directory.
+     */
+    cwd: string;
+    /**
+     * Error when copy failed.
+     */
+    error?: unknown;
+};
+
+/**
+ * Listener for copy end event.
+ */
+type CopyEndEventListener = (context: CopyEndEvent) => void | Promise<void>;
+
+export type {
+    CopyEndEvent,
+    CopyEndEventListener,
+    CopyEvent,
+    CopyEventListener,
+    CopyStartEvent,
+    CopyStartEventListener,
+    ResolvedTarget,
+    ResolvedTargetKind,
+};