+++ package/src/const/options.ts
@@ -3,12 +3,14 @@ import type { CompleteOptions } from "#/@types/options";
 const OPTIONS_DEFAULT = {
     cwd: process.cwd(),
     hook: "generateBundle",
+    targets: [],
     copyOnce: false,
     copySync: false,
     flatten: false,
     verbose: false,
+    onStart: void 0,
     onCopy: void 0,
-    targets: [],
+    onEnd: void 0,
 } as const satisfies CompleteOptions;
 
 export { OPTIONS_DEFAULT };